Add work_days_per_week, shared UI components, Toast system, and documentation
- Backend: Migration 006 for work_days_per_week field, updated schemas and calculation service
- Frontend: Shared components (Button, Badge, FormInput, FormSelect, FormTextarea, LoadingSpinner, MonthSelector, TableSkeleton)
- Frontend: Toast notification system (ToastContext/ToastProvider)
- Frontend: Responsive Layout improvements, updated all pages
- Docs: UX Roadmap, Benutzerhandbuch (V1 + V3), CLAUDE.md updates
- Scripts: Handbook screenshots, PDF generation, feature tests

+++ backend/app/schemas/working_hours_change.py
@@ -7,7 +7,7 @@ from uuid import UUID
 
 class WorkingHoursChangeBase(BaseModel):
     effective_from: date
-    weekly_hours: Decimal = Field(..., ge=0, le=60)
+    weekly_hours: float = Field(..., ge=0, le=60)
     note: Optional[str] = None
